fix(proxy): remove deprecated middleware.ts; add sales role enforcement to proxy.ts
Next.js 16 replaced the middleware file convention with proxy.ts.
The merge brought back the old middleware.ts from the feature branch.
Deleted it and merged the sales-role portal guard into proxy.ts.
